web_research_pipeline/requirements.txt
# Web Research Pipeline Dependencies
# Core dependencies for web search functionality

# Environment management
python-dotenv>=1.0.0

# HTTP requests for API calls
requests>=2.25.0

# Data processing and analysis
numpy>=1.21.0
pandas>=1.3.0

# Vector database and embeddings
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0

# Machine learning and model handling
torch>=1.9.0
transformers>=4.20.0

# Web scraping and content processing
beautifulsoup4>=4.10.0
lxml>=4.6.0

# PDF processing
PyMuPDF>=1.19.0

# Image processing
Pillow>=8.3.0
pytesseract>=0.3.8

# Logging and monitoring
loguru>=0.6.0

# Caching
diskcache>=5.2.0

# Rate limiting
ratelimit>=2.2.1

# Data validation
pydantic>=1.9.0

# Scheduling for research analysis
schedule>=1.2.0

# Async support (optional)
aiohttp>=3.8.0
asyncio-throttle>=1.0.0

# Development and testing
pytest>=6.2.0
pytest-asyncio>=0.15.0
black>=21.0.0
flake8>=3.9.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0

# Optional: Advanced features
# Uncomment if you need these features

# OpenAI integration for advanced query generation
# openai>=0.27.0

# Google Cloud Vision for image analysis
# google-cloud-vision>=2.4.0

# AWS services for cloud deployment
# boto3>=1.26.0

# Redis for distributed caching
# redis>=4.0.0

# PostgreSQL for persistent storage
# psycopg2-binary>=2.9.0

# MongoDB for document storage
# pymongo>=4.0.0

# Elasticsearch for advanced search
# elasticsearch>=8.0.0

# Prometheus for metrics
# prometheus-client>=0.14.0

# Grafana for monitoring
# grafana-api>=1.0.0 